fix: align serving with migration contracts
- enforce plan-owned environment and endpoint behavior
- limit compatibility to tested vLLM release series
- split shared and vLLM configuration from run records
- cover multi-group ordering and launch delays

Refs #866
