+++ packages/data-designer-slurm/src/data_designer/slurm/serving/compatibility.py
@@ -36,7 +36,6 @@ _V1_CAPABILITIES = _VllmCapabilities(
 _SUPPORTED_VLLM_CAPABILITIES = {
     (0, 21): _V1_CAPABILITIES,
     (0, 22): _V1_CAPABILITIES,
-    (0, 23): _V1_CAPABILITIES,
 }
