+++ packages/data-designer-slurm/tests/serving/test_compatibility.py
@@ -8,11 +8,11 @@ import pytest
 from data_designer.slurm.serving import UnsupportedServingRuntimeError, resolve_vllm_compatibility
 
 
-@pytest.mark.parametrize("runtime_version", ["0.21.0", "0.22.4", "0.23.0+container.1"])
+@pytest.mark.parametrize("runtime_version", ["0.21.0", "0.22.4+container.1"])
 def test_supported_vllm_release_series_have_complete_v1_capabilities(runtime_version: str) -> None:
     compatibility = resolve_vllm_compatibility(runtime_version)
 
-    assert compatibility.runtime_series in {"0.21", "0.22", "0.23"}
+    assert compatibility.runtime_series in {"0.21", "0.22"}
     assert compatibility.supports_single_node
     assert compatibility.supports_multi_node
     assert compatibility.supports_http_readiness
@@ -29,6 +29,7 @@ def test_supported_vllm_release_series_have_complete_v1_capabilities(runtime_ver
         "v0.21.0",
         "0.20.0",
         "0.21.0rc1",
+        "0.23.0+container.1",
         "0.24.0",
         "1!0.21.0",
         f"0.21.0+{'a' * 129}",
